Annotate dep modules
