+++ test/hello_world/gen/src/other.erl
@@ -1,6 +1,7 @@
 -module(other).
 -compile(no_auto_import).
 
+<<<<<<< HEAD
 solve_b(Noun) ->
     case Noun of
         99 ->
@@ -9,4 +10,10 @@ solve_b(Noun) ->
 
         Code ->
             Noun1
+=======
+run() ->
+    case {1, 2.0} of
+        {_, _} ->
+            1
+>>>>>>> Annotate dep modules
     end.
